Final Project
This project builds the hierarchical structure for the multi-resolution approximation in C++ using MPI. These files contain the code needed to run the model, however the Data folder and the Eigen library are not included. This code makes extensive use of the Eigen library for linear algebra purposes. To use the code, one must have Eigen installed. For more information go to http://eigen.tuxfamily.org/